Move grading functions to their own file
