+++ src/app.py
@@ -14,7 +14,8 @@ from flask_executor import Executor
 from flask_cors import CORS
 from concurrent.futures import ThreadPoolExecutor
 from app import create_app, start_grading_server, run_startup_scripts
-from app.functions import read_config, done_grading, get_logs
+from app.functions import read_config, get_logs
+from app.grading import done_grading
 from app.portServiceChecker import waitForService, checkServiceLoop, checkLocalPortLoop
 from app.extensions import logger, globals
